{{cookiecutter.project_slug}}/requirements.txt
django==2.2.8
django-db-geventpool
django-debug-toolbar
django-environ
django-hijack
django-htmlmin
django-parler
gevent
isort
pillow
psycopg2-binary
python-memcached
whitenoise
